move fixtures around to fix tests
