+++ src/lando/conftest.py
@@ -16,6 +16,7 @@ from lando.api.legacy.stacks import (
     request_extended_revision_data,
 )
 from lando.api.legacy.transplants import build_stack_assessment_state
+from lando.api.tests.mocks import TreeStatusDouble
 from lando.headless_api.models.tokens import ApiToken
 from lando.main.models import (
     SCM_LEVEL_1,
@@ -98,18 +99,14 @@ def normal_patch():
 
 
 @pytest.fixture
-def git_setup_user():
-    return _git_setup_user
+def git_repo_seed() -> Path:
+    """
+    Return the path to a patch to set up a base git repo for tests.
 
-
-def _git_setup_user(repo_dir: Path):
-    """Configure the git user locally to repo_dir so as not to mess with the real user's configuration."""
-    subprocess.run(["git", "config", "user.name", "Py Test"], check=True, cwd=repo_dir)
-    subprocess.run(
-        ["git", "config", "user.email", "pytest@lando.example.net"],
-        check=True,
-        cwd=repo_dir,
-    )
+    The diff can  apply on an empty repo to create a known base for application
+    of other patches as part of the tests.
+    """
+    return Path(__file__).parent / "main" / "tests" / "data"
 
 
 def _git_ignore_denyCurrentBranch(repo_dir: Path):
@@ -127,14 +124,24 @@ def _git_ignore_denyCurrentBranch(repo_dir: Path):
 
 
 @pytest.fixture
-def git_repo_seed() -> Path:
-    """
-    Return the path to a patch to set up a base git repo for tests.
+def git_setup_user():
+    return _git_setup_user
 
-    The diff can  apply on an empty repo to create a known base for application
-    of other patches as part of the tests.
-    """
-    return Path(__file__).parent / "main" / "tests" / "data" / "test-repo.patch"
+
+def _git_setup_user(repo_dir: Path):
+    """Configure the git user locally to repo_dir so as not to mess with the real user's configuration."""
+    _run_commands(
+        [
+            ["git", "config", "user.name", "Py Test"],
+            ["git", "config", "user.email", "pytest@lando.example.net"],
+        ],
+        repo_dir,
+    )
+
+
+def _run_commands(commands: list[list[str]], cwd: Path):
+    for c in commands:
+        subprocess.run(c, check=True, cwd=cwd)
 
 
 @pytest.fixture
@@ -153,15 +160,18 @@ def git_repo(tmp_path: Path, git_repo_seed: Path) -> Path:
     subprocess.run(["git", "branch", "-m", "main"], check=True, cwd=repo_dir)
     _git_setup_user(repo_dir)
     _git_ignore_denyCurrentBranch(repo_dir)
-    subprocess.run(["git", "am", str(git_repo_seed)], check=True, cwd=repo_dir)
+    for patch in sorted(git_repo_seed.glob("*")):
+        subprocess.run(["git", "am", str(patch)], check=True, cwd=repo_dir)
 
     # Create a separate base branch for branch tests.
-    subprocess.run(["git", "checkout", "-b", "dev"], check=True, cwd=repo_dir)
-    subprocess.run(
-        ["git", "commit", "--allow-empty", "-m", "dev"], check=True, cwd=repo_dir
+    _run_commands(
+        [
+            ["git", "checkout", "-b", "dev"],
+            ["git", "commit", "--allow-empty", "-m", "dev"],
+            ["git", "checkout", "main"],
+        ],
+        repo_dir,
     )
-
-    subprocess.run(["git", "checkout", "main"], check=True, cwd=repo_dir)
     return repo_dir
 
 
@@ -464,3 +474,15 @@ def create_state(
         )
 
     return create_state_handler
+
+
+@pytest.fixture
+def treestatus_url():
+    """A string holding the Tree Status base URL."""
+    return "http://treestatus.test"
+
+
+@pytest.fixture
+def treestatusdouble(monkeypatch, treestatus_url):
+    """Mock the Tree Status service and build fake responses."""
+    yield TreeStatusDouble(monkeypatch, treestatus_url)
